Minor BugFixes, Refactoring + Packaging.
