+++ app/src/main/java/picklenostra/picklebankapp/ProfileActivity.java
@@ -4,10 +4,12 @@ import android.os.Bundle;
 import android.support.design.widget.TabLayout;
 import android.support.v4.view.ViewPager;
 import android.support.v7.app.AppCompatActivity;
-import android.support.v7.app.ActionBar;
 import android.support.v7.widget.Toolbar;
 import android.view.Menu;
 import android.view.MenuItem;
+
+import picklenostra.picklebankapp.Adapter.PagerAdapter;
+import picklenostra.picklebankapp.Helper.UserSessionManager;
 
 
 public class ProfileActivity extends AppCompatActivity {
