+++ app/src/main/java/picklenostra/picklebankapp/Splash.java
@@ -4,8 +4,8 @@ import android.content.Intent;
 import android.os.Bundle;
 import android.os.Handler;
 import android.support.v7.app.ActionBarActivity;
-import android.view.Menu;
-import android.view.MenuItem;
+
+import picklenostra.picklebankapp.Helper.UserSessionManager;
 
 public class Splash extends ActionBarActivity {
