+++ app/src/main/res/layout/nasabah_listviewitem.xml
@@ -18,6 +18,7 @@
             android:id="@+id/nasabah_photo"
             android:layout_width="50dp"
             android:layout_height="50dp"
+            android:layout_marginLeft="8dp"
             android:src="@mipmap/ic_user"/>
     </LinearLayout>
